BL-1818 Move Book Covers js to Stimulus controller
